+++ app/src/main/java/com/example/task23/data/database/WeatherDatabase.kt
@@ -4,29 +4,10 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.task23.App
 
 @Database(entities = [WeatherEntity::class], version = 1, exportSchema = false)
 abstract class WeatherDatabase: RoomDatabase() {
 
     abstract fun weatherDao(): WeatherDao
-
-    companion object {
-        private var inst: WeatherDatabase? = null
-
-        fun getDatabase(context: Context): WeatherDatabase {
-            val tempInstance = inst
-            if (tempInstance != null) {
-                return tempInstance
-            }
-            synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    WeatherDatabase::class.java,
-                    "task_database"
-                ).build()
-                inst = instance
-                return instance
-            }
-        }
-    }
 }
